gumbo-dao/pyproject.toml
[tool.poetry]
name = "gumbo-dao"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "gumbo_dao"}]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9.9"
pandas = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
